# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IdentitySvc API",
        "description": "Vendor-agnostic identity facade over Azure AD B2C, Auth0, Okta, or Keycloak.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Start OIDC login",
                "description": "Returns an authorization URL (and PKCE challenge when applicable) that the client can use to redirect the user to the upstream identity provider.",
                "operationId": "33b4e968103a1b9265ea7346c45b3c1d",
                "parameters": [
                    {
                        "name": "pkce",
                        "in": "query",
                        "description": "Explicitly enable or disable PKCE for this request (overrides default behavior).",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": true
                    },
                    {
                        "name": "use_pkce",
                        "in": "query",
                        "description": "Alias for the pkce parameter.",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "client_type",
                        "in": "query",
                        "description": "Hint used to infer PKCE behavior. Common values: 'spa', 'native', 'server'.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "spa"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authorize URL payload",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "authorize_url": {
                                            "type": "string",
                                            "example": "https://login.identity.example.com/oauth2/v2.0/authorize?client_id=...&response_type=code&state=..."
                                        },
                                        "uses_pkce": {
                                            "description": "Echo of the resolved PKCE decision for this request",
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            }
        },
        "/api/auth/callback": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "OIDC callback (authorization code exchange)",
                "description": "Exchanges the authorization code for tokens. If PKCE was used for this 'state', the stored verifier is automatically applied.",
                "operationId": "9e77cef6c23d0c1e57c30a916d124651",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OIDC token response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tokens"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh tokens",
                "description": "Use a refresh token to obtain a fresh access token (and optionally new refresh/id tokens).",
                "operationId": "b2eae7a8032954d25b46c7ec6f489c68",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "refresh_token"
                                ],
                                "properties": {
                                    "refresh_token": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OIDC token response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tokens"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get current user profile",
                "description": "Returns the normalized user profile for the bearer access token supplied in the Authorization header.",
                "operationId": "9d9bc556f927575aa492a88c28fe5908",
                "responses": {
                    "200": {
                        "description": "User profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserProfile"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Performs local logout (stateless). Returns 204. The client should also sign out from the upstream identity provider if desired.",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Tokens": {
                "required": [
                    "expiresIn"
                ],
                "properties": {
                    "accessToken": {
                        "description": "Access token issued by the identity provider",
                        "type": "string",
                        "nullable": true
                    },
                    "refreshToken": {
                        "description": "Refresh token that can be used to obtain new access tokens",
                        "type": "string",
                        "nullable": true
                    },
                    "expiresIn": {
                        "description": "Lifetime of the access token in seconds",
                        "type": "integer",
                        "example": 3600
                    },
                    "idToken": {
                        "description": "ID token (when available)",
                        "type": "string",
                        "nullable": true
                    },
                    "tokenType": {
                        "description": "Token type as reported by the provider",
                        "type": "string",
                        "example": "Bearer",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UserProfile": {
                "required": [
                    "id",
                    "email",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "givenName": {
                        "type": "string",
                        "nullable": true
                    },
                    "familyName": {
                        "type": "string",
                        "nullable": true
                    },
                    "displayName": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "invited"
                        ]
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "attributes": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "BadRequest": {
                "description": "Bad Request",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "details": {
                                    "type": "object",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "details": {
                                    "type": "object",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "TooManyRequests": {
                "description": "Too Many Requests",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "details": {
                                    "type": "object",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ServerError": {
                "description": "Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "details": {
                                    "type": "object",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
...
}